Broken up the GlobalEvents file and refactored in some cases
