+++ indigo-core/src/indigo/core/events/PointerType.scala
@@ -1,0 +1,6 @@
+package indigo.core.events
+
+/** The type of pointer that has emitted an input pointer event
+  */
+enum PointerType derives CanEqual:
+  case Mouse, Pen, Touch, Unknown
