+++ indigo-core/src/indigo/core/events/SubSystemEvent.scala
@@ -1,0 +1,5 @@
+package indigo.core.events
+
+/** A trait whose presence signals that this event should only be routed to subsystems, not the main game.
+  */
+trait SubSystemEvent extends GlobalEvent
