+++ indigo-core/src/indigo/core/events/WheelDirection.scala
@@ -1,0 +1,6 @@
+package indigo.core.events
+
+/** Represents in which direction the wheel input was rotated
+  */
+enum WheelDirection derives CanEqual:
+  case Up, Down, Left, Right
